Calculating insurance discounts
Write a program which calculates the insurance premium for the coming year. The program asks the user for the value of last year’s premium and the number of claims you made in the year. The basic premium for the following year  will be a 5% increase on last years premium. If the person made no claims they get a 40% discount on this increase. If they made a claim the get an extra charge of 20% for every claim that they made.